fix(数据迁移): 会员卡有效期

+++ main/app/model/member.go
@@ -240,7 +240,7 @@ func (model *MemberCard) GetDiscount() float64 {
 type MemberCardType struct {
 	BaseModel
 	Name         string  `gorm:"column:name;type:varchar(255);comment:会员卡类型名称;NOT NULL" json:"name"`
-	Expire       int     `gorm:"column:expire;type:int(11);default:0;comment:有效期限,单位:月, 0为永久有效;NOT NULL" json:"expire"`
+	Expire       int64   `gorm:"column:expire;type:int(11);default:0;comment:有效期限,单位:月, 0为永久有效;NOT NULL" json:"expire"`
 	Price        float64 `gorm:"column:price;type:decimal(12,2);default:0.00;comment:价格;NOT NULL" json:"price"`
 	Discount     float64 `gorm:"column:discount;type:tinyint(3);default:0;comment:折扣,单位%;NOT NULL" json:"discount"`
 	Sort         int     `gorm:"column:sort;type:int(11);default:0;comment:排序;NOT NULL" json:"sort"`
@@ -257,7 +257,7 @@ type MemberCardLog struct {
 	BaseModel
 	Price              float64 `gorm:"column:price;type:decimal(12,2);default:0.00;comment:价格,会员卡价格,不随后台改变,记录领取时的价格;NOT NULL" json:"price"`
 	Discount           float64 `gorm:"column:discount;type:tinyint(3);default:0;comment:折扣,单位%,不随后台改变,记录领取时的折扣;NOT NULL" json:"discount"`
-	Expire             int     `gorm:"column:expire;type:int(11);default:0;comment:有效期限,单位:月, 0为永久有效,不随后台改变,记录领取时的有效期限;NOT NULL" json:"expire"`
+	Expire             int64   `gorm:"column:expire;type:int(11);default:0;comment:有效期限,单位:月, 0为永久有效,不随后台改变,记录领取时的有效期限;NOT NULL" json:"expire"`
 	MemberName         string  `gorm:"column:member_name;type:varchar(255);comment:会员名称,不随后台改变,当无法用member_uuid获取会员信息时,用此字段;NOT NULL" json:"member_name"`
 	MemberPhone        string  `gorm:"column:member_phone;type:varchar(255);comment:会员电话,不随后台改变,当无法用member_uuid获取会员信息时,用此字段;NOT NULL" json:"member_phone"`
 	MemberNo           string  `gorm:"column:member_no;type:varchar(255);comment:会员编号,不随后台改变,当无法用member_uuid获取会员信息时,用此字段;NOT NULL" json:"member_no"`
